+++ Malfunctions/Patches/ManualCameraRendererPatches.cs
@@ -9,8 +9,11 @@ namespace Malfunctions.Patches
         [HarmonyPatch("SwitchScreenButton")]
         private static bool InterruptSwitchScreenButton()
         {
-            // Interrupt the screen button if a power malfunction has triggered.
-            if (State.MalfunctionPower.Active && State.MalfunctionPower.Triggered)
+            // Interrupt the screen button if a power or distortion malfunction has triggered.
+            if (
+                State.MalfunctionPower.Active && State.MalfunctionPower.Triggered
+                || State.MalfunctionDistortion.Active && State.MalfunctionDistortion.Triggered
+            )
             {
                 return false;
             }
@@ -22,8 +25,11 @@ namespace Malfunctions.Patches
         [HarmonyPatch("SwitchRadarTargetClientRpc")]
         private static bool InterruptSwitchCameraView()
         {
-            // Interrupt the screen button if a power malfunction has triggered.
-            if (State.MalfunctionPower.Active && State.MalfunctionPower.Triggered)
+            // Interrupt the screen button if a power or distortion malfunction has triggered.
+            if (
+                State.MalfunctionPower.Active && State.MalfunctionPower.Triggered
+                || State.MalfunctionDistortion.Active && State.MalfunctionDistortion.Triggered
+            )
             {
                 return false;
             }